Port Fix player_examine_nearby_unit: correct addr 0xbae10->0xbae20 + activate + int weapon_count (players.obj) (79.3% VC71)
Functions ported:
- player_examine_nearby_unit @ 0xbae20 (players.obj) [79.3% VC71, 61/60 insns]

Coverage: 30.5% (2479/8141 symbols)

+++ src/halo/game/players.c
@@ -572,7 +572,7 @@ bool player_examine_nearby_unit(int player_unit_handle, int nearby_unit_handle)
 {
   int *nearby_obj;
   char *weap_tag;
-  short weapon_count;
+  int weapon_count;
   bool can_swap;
 
   nearby_obj = (int *)object_try_and_get_and_verify_type(nearby_unit_handle, 4);
